fix(cli): Clear current input on ctrl-c instead of exiting
Most readline-style programs allow the user to clear the current line by
using `ctrl-c`, prior to this change `ctrl-c` would immediately exit
goose.

This change makes it so that `ctrl-c` just empties the input and puts
the user on an empty prompt line. The user can sill exit by using
`ctrl-d` or `/exit` or `/quit`.