feat(developer_shell): Add `output_tmpfile` parameter to shell tool
The goal is to lower the amount of tokens the LLM needs to parse for large
commands like `make` and `cargo build`. Most of the time they only need
stderr if they fail, and if they *don't* fail then there's no need to parse
the output at all usually.

As noted in the text this takes an optional regular expression and for
key tasks like `cargo build` I think a regex filter of `^(warning)|(error):`
works well here, avoiding a latency hit where the LLM may need
to check the redirected output file for warnings.

I also changed things so that we don't have a hardcoded command output
limit, but instead we just automatically redirect to a tempfile
anyways if it's large.

Finally as part of the implmentation of this I cleaned up the way
we read from stdout/stderr to greatly deduplicate code by iterating
over a stream.

Signed-off-by: Colin Walters <walters@verbum.org>