chore(mcp): Add helpers to parse parameters
- Avoids redundant code
- Doesn't ignore parameters on type mismatches

Motivated specifically by me having a typo in my optional parameter name
when developing.

Note: I only changed a random selection of goose-mcp, but if
we choose to do this it'd likely be trivial to ask goose to
finish it. I didn't do so yet to not make this a conflict-fest.

Signed-off-by: Colin Walters <walters@verbum.org>