Add a setting for the quit confirmation dialog
This adds a setting to allow the user to enable or disable the quit confimation
dialog. The setting is enabled by default (to keep the current behaviour).

The reason I'm adding this, is that I find these quit confirmations annoying.
I shut down my computer each night, and this dialog prevents shutdown until I hit
OK.