mcp/developer: Refactor to use tokio SplitStream
I was working on a PR with a lot larger output set sizes
and I *think* I was seeing the last chunk of output sometimes being
lost.

That said I asked Gemini to try to reproduce the failure
with a case and I couldn't get it to fail offhand. In looking
at the original code here it looks OK...

Async rust is hard; `tokio::select!` in particular can cause
subtle bugs. For a lot more on this see
https://blog.yoshuawuyts.com/futures-concurrency-3
and various previous and future blog entries.

Anyways this code is I think cleaner, deduplicating the code
to send the notification message.

Signed-off-by: Colin Walters <walters@verbum.org>