Fix bedrock tool call error response
When a tool call fails, bedrock expects that a response is returned back
to the api regardless of the tool call result. Prior to this patch,
bedrock would throw an exception that goose could not handle. This patch
fixes this by returning a text content block with the error message
instead of returning nothing.

Signed-off-by: alexyao2015 <alexyao2015@users.noreply.github.com>