Refactor GitHub PR Summarizer configuration
I removed the 'summarizer' MCP server, since it requires some external setup (cloning the repo and then adding it as an extension manually to point at the correct file. Most LLMs are going to do a fine job building summary text without needing a third-party MCP server for this task.

I added a prompt for headless mode goose as well, and included a github_repository parameter to use as well.

I also included the streamable-http GitHub MCP server as well, and included a uri for enterprise and non-enterprise users.