Fix `nix run` (nix-flake mainProgram).
Running `nix run` on a clean checkout would build and then fail because
`goose-cli` could not be found. It can't be found because it's actually
named just `goose`.

This fixes running `nix run github:block/goose` or `nix run` on a local
copy of the source. Nix flake defaults to running `goose-cli`, this
patch changes that to `goose` (which matches the built binary).

Signed-off-by: Henrik Sjöstrand <henrik.sjostrand@tele2.com>